Rename assistant module to agent
